refactor(core): move SQLDelight dao in mobileMain sourceset.
