+++ shared/core/src/commonMain/kotlin/com/paligot/confily/core/networking/NetworkingRepository.kt
@@ -4,6 +4,7 @@ import com.paligot.confily.core.QrCodeGenerator
 import com.paligot.confily.core.kvalue.ConferenceSettings
 import com.paligot.confily.core.extensions.encodeToString
 import com.paligot.confily.core.fs.ConferenceFileSystem
+import com.paligot.confily.core.fs.User
 import com.paligot.confily.models.ui.ExportNetworkingUi
 import com.paligot.confily.models.ui.UserNetworkingUi
 import com.paligot.confily.models.ui.UserProfileUi
@@ -93,3 +94,10 @@ class NetworkingRepositoryImpl(
         )
     }
 }
+
+private fun UserNetworkingUi.toUser(): User = User(
+    email = email,
+    firstname = firstName,
+    lastname = lastName,
+    company = company ?: ""
+)
