+++ shared/core/src/mobileMain/kotlin/com/paligot/confily/core/speakers/DbMappers.kt
